+++ lib/aemo/nem12.rb
@@ -297,9 +297,7 @@ module AEMO
 
       raise ArgumentError, 'RecordIndicator is not 100'     if csv[0] != '100'
       raise ArgumentError, 'VersionHeader is not NEM12'     if csv[1] != 'NEM12'
-      if options[:strict] && (csv[2].match(/\d{12}/).nil? || csv[2] != Time.parse("#{csv[2]}00").strftime('%Y%m%d%H%M'))
-        raise ArgumentError, 'DateTime is not valid'
-      end
+      raise ArgumentError, 'Time is not valid' if options[:strict] && (csv[2].match(/\d{12}/).nil? || csv[2] != Time.parse("#{csv[2]}00").strftime('%Y%m%d%H%M'))
       raise ArgumentError, 'FromParticipant is not valid'  if csv[3].match(/.{1,10}/).nil?
       raise ArgumentError, 'ToParticipant is not valid'    if csv[4].match(/.{1,10}/).nil?
 
@@ -320,24 +318,12 @@ module AEMO
 
       raise ArgumentError, 'RecordIndicator is not 200'     if csv[0] != '200'
       raise ArgumentError, 'NMI is not valid'               unless AEMO::NMI.valid_nmi?(csv[1])
-      if options[:strict] && (csv[2].nil? || csv[2].match(/.{1,240}/).nil?)
-        raise ArgumentError, 'NMIConfiguration is not valid'
-      end
-      if !csv[3].nil? && csv[3].match(/.{1,10}/).nil?
-        raise ArgumentError, 'RegisterID is not valid'
-      end
-      if csv[4].nil? || csv[4].match(/[A-HJ-NP-Z][1-9A-HJ-NP-Z]/).nil?
-        raise ArgumentError, 'NMISuffix is not valid'
-      end
-      if !csv[5].nil? && !csv[5].empty? && !csv[5].match(/^\s*$/)
-        raise ArgumentError, 'MDMDataStreamIdentifier is not valid' if csv[5].match(/[A-Z0-9]{2}/).nil?
-      end
-      if !csv[6].nil? && !csv[6].empty? && !csv[6].match(/^\s*$/)
-        raise ArgumentError, 'MeterSerialNumber is not valid' if csv[6].match(/[A-Z0-9]{2}/).nil?
-      end
-      if csv[7].nil? || csv[7].upcase.match(/[A-Z0-9]{2}/).nil?
-        raise ArgumentError, 'UOM is not valid'
-      end
+      raise ArgumentError, 'NMIConfiguration is not valid' if options[:strict] && (csv[2].nil? || csv[2].match(/.{1,240}/).nil?)
+      raise ArgumentError, 'RegisterID is not valid' if !csv[3].nil? && csv[3].match(/.{1,10}/).nil?
+      raise ArgumentError, 'NMISuffix is not valid' if csv[4].nil? || csv[4].match(/[A-HJ-NP-Z][1-9A-HJ-NP-Z]/).nil?
+      raise ArgumentError, 'MDMDataStreamIdentifier is not valid' if !csv[5].nil? && !csv[5].empty? && !csv[5].match(/^\s*$/) && csv[5].match(/[A-Z0-9]{2}/).nil?
+      raise ArgumentError, 'MeterSerialNumber is not valid' if !csv[6].nil? && !csv[6].empty? && !csv[6].match(/^\s*$/) && csv[6].match(/[A-Z0-9]{2}/).nil?
+      raise ArgumentError, 'UOM is not valid' if csv[7].nil? || csv[7].upcase.match(/[A-Z0-9]{2}/).nil?
       raise ArgumentError, 'UOM is not valid'               unless UOM.keys.map(&:upcase).include?(csv[7].upcase)
       raise ArgumentError, 'IntervalLength is not valid'    unless %w[1 5 10 15 30].include?(csv[8])
       # raise ArgumentError, 'NextScheduledReadDate is not valid' if csv[9].match(/\d{8}/).nil? || csv[9] != Time.parse('#{csv[9]}').strftime('%Y%m%d')
@@ -404,9 +390,7 @@ module AEMO
           flag[:quality_flag] = csv[intervals_offset + 0][0]
           flag[:method_flag] = csv[intervals_offset + 0][1, 2].to_i
         end
-        unless csv[intervals_offset + 1].nil?
-          flag[:reason_code] = csv[intervals_offset + 1].to_i
-        end
+        flag[:reason_code] = csv[intervals_offset + 1].to_i unless csv[intervals_offset + 1].nil?
       end
 
       base_interval = { data_details: @data_details.last, datetime: Time.parse("#{csv[1]}000000+1000"), value: nil, flag: flag }
@@ -454,9 +438,7 @@ module AEMO
             flag[:quality_flag] = interval_event[:quality_method][0]
             flag[:method_flag] = interval_event[:quality_method][1, 2].to_i
           end
-          unless interval_event[:reason_code].nil?
-            flag[:reason_code] = interval_event[:reason_code]
-          end
+          flag[:reason_code] = interval_event[:reason_code] unless interval_event[:reason_code].nil?
           # Update with flag details
           @interval_data[interval_start_point + (i - 1)][:flag] = flag
         end